Function ispc::compile_library

pub fn compile_library(lib: &str, files: &[&str])

Compile the list of ISPC files into a static library and generate bindings using bindgen. The library name should not contain a lib prefix or a lib extension like '.a' or '.lib', the appropriate prefix and suffix will be added based on the compilation target.

This function will exit the process with EXIT_FAILURE if any stage of compilation or linking fails.

Example

ispc::compile_library("foo", &["src/foo.ispc", "src/bar.ispc"]);