Module ispc::opt

This module has various option flags and configs we can pass to ISPC, located here for convience and clutter reduction.

Enums

Addressing
Select 32 or 64 bit addressing to be used by ISPC. Note: 32-bit addressing calculations are done by default, even on 64 bit target architectures.

CPU
ISPC target CPU type options, if none is set ISPC will target the machine being compile on.

MathLib
Different math libraries that ISPC can use for computations.

OptimizationOpt
ISPC optimization options.

TargetISA
Target instruction sets and vector widths available to specialize for. The default if none is set will be the host CPU's ISA and vector width.