Enum ispc::opt::TargetISA

pub enum TargetISA {
    Host,
    SSE2i32x4,
    SSE2i32x8,
    SSE4i32x4,
    SSE4i32x8,
    SSE4i16x8,
    SSE4i8x16,
    AVX1i32x4,
    AVX1i32x8,
    AVX1i32x16,
    AVX1i64x4,
    AVX11i32x8,
    AVX11i32x16,
    AVX11i64x4,
    AVX2i32x8,
    AVX2i32x16,
    AVX2i64x4,
    AVX512KNLi32x16,
}

Target instruction sets and vector widths available to specialize for. The default if none is set will be the host CPU's ISA and vector width.

Variants
HostSSE2i32x4SSE2i32x8SSE4i32x4SSE4i32x8SSE4i16x8SSE4i8x16AVX1i32x4AVX1i32x8AVX1i32x16AVX1i64x4AVX11i32x8AVX11i32x16AVX11i64x4AVX2i32x8AVX2i32x16AVX2i64x4AVX512KNLi32x16
Methods

impl TargetISA

pub fn lib_suffix(&self) -> String