Function ispc::set_task_system

pub fn set_task_system<F>(f: F) where
    F: FnOnce() -> Arc<dyn TaskSystem + 'static>, 

If you have implemented your own task system you can provide it for use instead of the default threaded one. This must be done prior to calling ISPC code which spawns tasks otherwise the task system will have already been initialized to Parallel, which you can also see as an example for implementing a task system.

Use the function to do any extra initialization for your task system. Note that the task system will be leaked and not destroyed until the program exits and the memory space is cleaned up.